Fix taplist projection migration for Postgres

+++ src/main/resources/db/migration/V5__taplist_projections.sql
@@ -5,12 +5,11 @@ create table if not exists taplist_view (
   venue_id bigint,
   beer_name varchar(255),
   style varchar(255),
-  abv double,
-  remaining_ounces double,
-  total_ounces double,
+  abv double precision,
+  remaining_ounces double precision,
+  total_ounces double precision,
   fill_percent integer,
   updated_at timestamp not null
 );
 
 create index if not exists idx_tlv_venue on taplist_view(venue_id);
-
